Revert from bigint to int

+++ prisma/migrations/20251104120043_init/migration.sql
@@ -9,7 +9,7 @@ CREATE TYPE "PurchaseOrderStatus" AS ENUM ('SUBMITTED', 'PAID', 'ORDERED_FROM_VE
 
 -- CreateTable
 CREATE TABLE "User" (
-    "id" BIGSERIAL NOT NULL,
+    "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
     "phoneNumber" TEXT NOT NULL,
     "otp" INTEGER NOT NULL,
@@ -27,9 +27,9 @@ CREATE TABLE "User" (
 
 -- CreateTable
 CREATE TABLE "Login" (
-    "id" BIGSERIAL NOT NULL,
+    "id" SERIAL NOT NULL,
     "ipAddress" INET,
-    "userId" BIGINT NOT NULL,
+    "userId" INTEGER NOT NULL,
     "tokenValue" TEXT NOT NULL,
     "explicit" BOOLEAN NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -40,7 +40,7 @@ CREATE TABLE "Login" (
 -- CreateTable
 CREATE TABLE "AccessToken" (
     "value" TEXT NOT NULL,
-    "userId" BIGINT NOT NULL,
+    "userId" INTEGER NOT NULL,
     "expired" BOOLEAN NOT NULL DEFAULT false,
 
     CONSTRAINT "AccessToken_pkey" PRIMARY KEY ("value")
@@ -48,7 +48,7 @@ CREATE TABLE "AccessToken" (
 
 -- CreateTable
 CREATE TABLE "Product" (
-    "id" BIGSERIAL NOT NULL,
+    "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
     "upc" TEXT,
     "upcScanned" BOOLEAN,
@@ -72,12 +72,12 @@ CREATE TABLE "Product" (
 
 -- CreateTable
 CREATE TABLE "Item" (
-    "id" BIGSERIAL NOT NULL,
-    "userId" BIGINT NOT NULL,
+    "id" SERIAL NOT NULL,
+    "userId" INTEGER NOT NULL,
     "code" TEXT NOT NULL,
     "name" TEXT NOT NULL,
     "description" TEXT,
-    "productId" BIGINT,
+    "productId" INTEGER,
     "weightInKgs" DOUBLE PRECISION,
     "widthInCms" DOUBLE PRECISION,
     "heightInCms" DOUBLE PRECISION,
@@ -90,8 +90,8 @@ CREATE TABLE "Item" (
 
 -- CreateTable
 CREATE TABLE "PurchaseOrder" (
-    "id" BIGSERIAL NOT NULL,
-    "userId" BIGINT NOT NULL,
+    "id" SERIAL NOT NULL,
+    "userId" INTEGER NOT NULL,
     "price" MONEY NOT NULL,
     "status" "PurchaseOrderStatus" NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -102,10 +102,10 @@ CREATE TABLE "PurchaseOrder" (
 
 -- CreateTable
 CREATE TABLE "PurchaseOrderEntry" (
-    "id" BIGSERIAL NOT NULL,
-    "orderId" BIGINT NOT NULL,
+    "id" SERIAL NOT NULL,
+    "orderId" INTEGER NOT NULL,
     "quantity" INTEGER NOT NULL,
-    "productId" BIGINT NOT NULL,
+    "productId" INTEGER NOT NULL,
     "currency" "Currency" NOT NULL,
     "unitPrice" MONEY NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
